Tidy up nxs_write
