+++ src/nexgen/nxs_copy/CopyTristanNexus.py
@@ -9,8 +9,8 @@ from pathlib import Path
 import h5py
 import numpy as np
 
-from ..nxs_write import create_attributes
 from ..nxs_write.NXclassWriters import write_NXnote
+from ..nxs_write.write_utils import create_attributes
 from . import (
     check_and_fix_det_axis,
     compute_ssx_axes,
